- add OCSP stapling - upgrade to haproxytech/client-native/v6 - upgrade to go 1.23.7 - bump dependencies - add golangci-lint rules - bump github checks

+++ controllers/instance/route.go
@@ -63,7 +63,7 @@ func (r *Reconciler) createOrUpdateRouteForFrontend(ctx context.Context, instanc
 			}
 
 			route.Spec.Port = &routev1.RoutePort{
-				TargetPort: intstr.FromInt32(int32(bind.Port)),
+				TargetPort: intstr.FromInt32(bind.Port),
 			}
 
 			route.Spec.TLS = instance.Spec.Network.Route.TLS
